requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.8.0rc2
Django==5.0.3
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
google-auth==2.28.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.2
